feat(mysfire): Changed collate to flatten dictionaries at top level
The behavior of the collate function now flattens top level dictionaries using a '.' separator as a key. If the key in the second-level dictionary is __root__, the value will be placed in the original key value (making it easy to add processors that introduce additional top-level information, without affecting the core data returned). This is a breaking change for pretty much everything.